+++ pkg/api/query/v1.go
@@ -53,11 +53,11 @@ import (
 )
 
 const (
-	dedupParam               = "dedup"
-	partialResponseParam     = "partial_response"
-	maxSourceResolutionParam = "max_source_resolution"
-	replicaLabelsParam       = "replicaLabels[]"
-	storeMatcherParam        = "storeMatch[]"
+	DedupParam               = "dedup"
+	PartialResponseParam     = "partial_response"
+	MaxSourceResolutionParam = "max_source_resolution"
+	ReplicaLabelsParam       = "replicaLabels[]"
+	StoreMatcherParam        = "storeMatch[]"
 )
 
 // QueryAPI is an API used by Thanos Query.
@@ -149,11 +149,11 @@ type queryData struct {
 func (qapi *QueryAPI) parseEnableDedupParam(r *http.Request) (enableDeduplication bool, _ *api.ApiError) {
 	enableDeduplication = true
 
-	if val := r.FormValue(dedupParam); val != "" {
+	if val := r.FormValue(DedupParam); val != "" {
 		var err error
 		enableDeduplication, err = strconv.ParseBool(val)
 		if err != nil {
-			return false, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Wrapf(err, "'%s' parameter", dedupParam)}
+			return false, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Wrapf(err, "'%s' parameter", DedupParam)}
 		}
 	}
 	return enableDeduplication, nil
@@ -166,8 +166,8 @@ func (qapi *QueryAPI) parseReplicaLabelsParam(r *http.Request) (replicaLabels []
 
 	replicaLabels = qapi.replicaLabels
 	// Overwrite the cli flag when provided as a query parameter.
-	if len(r.Form[replicaLabelsParam]) > 0 {
-		replicaLabels = r.Form[replicaLabelsParam]
+	if len(r.Form[ReplicaLabelsParam]) > 0 {
+		replicaLabels = r.Form[ReplicaLabelsParam]
 	}
 
 	return replicaLabels, nil
@@ -178,7 +178,7 @@ func (qapi *QueryAPI) parseStoreMatchersParam(r *http.Request) (storeMatchers []
 		return nil, &api.ApiError{Typ: api.ErrorInternal, Err: errors.Wrap(err, "parse form")}
 	}
 
-	for _, s := range r.Form[storeMatcherParam] {
+	for _, s := range r.Form[StoreMatcherParam] {
 		matchers, err := parser.ParseMetricSelector(s)
 		if err != nil {
 			return nil, &api.ApiError{Typ: api.ErrorBadData, Err: err}
@@ -196,7 +196,7 @@ func (qapi *QueryAPI) parseStoreMatchersParam(r *http.Request) (storeMatchers []
 func (qapi *QueryAPI) parseDownsamplingParamMillis(r *http.Request, defaultVal time.Duration) (maxResolutionMillis int64, _ *api.ApiError) {
 	maxSourceResolution := 0 * time.Second
 
-	val := r.FormValue(maxSourceResolutionParam)
+	val := r.FormValue(MaxSourceResolutionParam)
 	if qapi.enableAutodownsampling || (val == "auto") {
 		maxSourceResolution = defaultVal
 	}
@@ -204,12 +204,12 @@ func (qapi *QueryAPI) parseDownsamplingParamMillis(r *http.Request, defaultVal t
 		var err error
 		maxSourceResolution, err = parseDuration(val)
 		if err != nil {
-			return 0, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Wrapf(err, "'%s' parameter", maxSourceResolutionParam)}
+			return 0, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Wrapf(err, "'%s' parameter", MaxSourceResolutionParam)}
 		}
 	}
 
 	if maxSourceResolution < 0 {
-		return 0, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Errorf("negative '%s' is not accepted. Try a positive integer", maxSourceResolutionParam)}
+		return 0, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Errorf("negative '%s' is not accepted. Try a positive integer", MaxSourceResolutionParam)}
 	}
 
 	return int64(maxSourceResolution / time.Millisecond), nil
@@ -217,11 +217,11 @@ func (qapi *QueryAPI) parseDownsamplingParamMillis(r *http.Request, defaultVal t
 
 func (qapi *QueryAPI) parsePartialResponseParam(r *http.Request, defaultEnablePartialResponse bool) (enablePartialResponse bool, _ *api.ApiError) {
 	// Overwrite the cli flag when provided as a query parameter.
-	if val := r.FormValue(partialResponseParam); val != "" {
+	if val := r.FormValue(PartialResponseParam); val != "" {
 		var err error
 		defaultEnablePartialResponse, err = strconv.ParseBool(val)
 		if err != nil {
-			return false, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Wrapf(err, "'%s' parameter", partialResponseParam)}
+			return false, &api.ApiError{Typ: api.ErrorBadData, Err: errors.Wrapf(err, "'%s' parameter", PartialResponseParam)}
 		}
 	}
 	return defaultEnablePartialResponse, nil