+++ pkg/api/query/v1_test.go
@@ -1056,7 +1056,7 @@ func TestParseDownsamplingParamMillis(t *testing.T) {
 			gate:                   gate.NewKeeper(nil).NewGate(4),
 		}
 		v := url.Values{}
-		v.Set(maxSourceResolutionParam, test.maxSourceResolutionParam)
+		v.Set(MaxSourceResolutionParam, test.maxSourceResolutionParam)
 		r := http.Request{PostForm: v}
 
 		// If no max_source_resolution is specified fit at least 5 samples between steps.
@@ -1099,19 +1099,22 @@ func TestParseStoreMatchersParam(t *testing.T) {
 			}},
 		},
 	} {
-		api := QueryAPI{
-			gate: gate.NewKeeper(nil).NewGate(4),
-		}
-		v := url.Values{}
-		v.Set(storeMatcherParam, tc.storeMatchers)
-		r := &http.Request{PostForm: v}
+		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
+			api := QueryAPI{
+				gate: gate.NewKeeper(nil).NewGate(4),
+			}
+			v := url.Values{}
+			v.Set(StoreMatcherParam, tc.storeMatchers)
+			r := &http.Request{PostForm: v}
 
-		storeMatchers, err := api.parseStoreMatchersParam(r)
-		if !tc.fail {
-			testutil.Assert(t, reflect.DeepEqual(storeMatchers, tc.result), "case %v: expected %v to be equal to %v", i, storeMatchers, tc.result)
-		} else {
-			testutil.NotOk(t, err)
-		}
+			storeMatchers, err := api.parseStoreMatchersParam(r)
+			if !tc.fail {
+				testutil.Equals(t, tc.result, storeMatchers)
+				testutil.Equals(t, (*baseAPI.ApiError)(nil), err)
+			} else {
+				testutil.NotOk(t, err)
+			}
+		})
 	}
 }